+++ CodeRed/Vulkan/VulkanUtility.hpp
@@ -97,6 +97,8 @@ namespace CodeRed {
 
 		auto enumConvert(const PixelFormat format, const ResourceUsage usage) -> vk::ImageAspectFlags;
 
+		auto enumConvert(const size_t arrayLength = 0) -> vk::ImageCreateFlags;
+		
 		auto enumConvert(const Dimension dimension, const TextureRefUsage type, const size_t depth) -> vk::ImageViewType;
 		
 		auto enumConvert1(const Dimension dimension)->vk::ImageViewType;